Fix warnings

+++ src/adpcm_inline.h
@@ -116,8 +116,8 @@ INLINE void Adpcm::Write(u16 address, u8 value)
 
 INLINE u32 Adpcm::CalculateCyclesPerSample(u8 sample_rate)
 {
-    float frequency = 32000.0f / (16.0f - (float)sample_rate);
-    return (u32)(GG_MASTER_CLOCK_RATE / frequency);
+    double frequency = 32000.0 / (16.0 - (double)sample_rate);
+    return (u32)((double)GG_MASTER_CLOCK_RATE / frequency);
 }
 
 INLINE u32 Adpcm::NextSlotCycles(bool read)
